+++ src/stitching/index.ts
@@ -1,4 +1,4 @@
-import makeRemoteExecutableSchema from './makeRemoteExecutableSchema';
+import makeRemoteExecutableSchema, { createResolver as defaultCreateRemoteResolver } from './makeRemoteExecutableSchema';
 import introspectSchema from './introspectSchema';
 import mergeSchemas from './mergeSchemas';
 import delegateToSchema, { createRequest } from './delegateToSchema';
@@ -12,5 +12,6 @@ export {
   // but exposed for the community use
   delegateToSchema,
   defaultMergedResolver,
-  createRequest
+  createRequest,
+  defaultCreateRemoteResolver
 };