+++ src/transforms/FilterRootFields.ts
@@ -1,4 +1,4 @@
-import { GraphQLField } from 'graphql';
+import { GraphQLField, GraphQLSchema } from 'graphql';
 import { Transform } from './transforms';
 import TransformRootFields from './TransformRootFields';
 
@@ -8,18 +8,26 @@ export type RootFilter = (
   field: GraphQLField<any, any>,
 ) => Boolean;
 
-export default function FilterRootFields(filter: RootFilter): Transform {
-  return TransformRootFields(
-    (
-      operation: 'Query' | 'Mutation' | 'Subscription',
-      fieldName: string,
-      field: GraphQLField<any, any>,
-    ) => {
-      if (filter(operation, fieldName, field)) {
-        return undefined;
-      } else {
-        return null;
-      }
-    },
-  );
+export default class FilterRootFields implements Transform {
+  private transformer: TransformRootFields;
+
+  constructor(filter: RootFilter) {
+    this.transformer = new TransformRootFields(
+      (
+        operation: 'Query' | 'Mutation' | 'Subscription',
+        fieldName: string,
+        field: GraphQLField<any, any>,
+      ) => {
+        if (filter(operation, fieldName, field)) {
+          return undefined;
+        } else {
+          return null;
+        }
+      },
+    );
+  }
+
+  public transformSchema(originalSchema: GraphQLSchema): GraphQLSchema {
+    return this.transformer.transformSchema(originalSchema);
+  }
 }