+++ src/transforms/ReplaceFieldWithFragment.ts
@@ -6,34 +6,51 @@ import {
   Kind,
   SelectionSetNode,
   TypeInfo,
+  OperationDefinitionNode,
+  parse,
   visit,
   visitWithTypeInfo,
 } from 'graphql';
 import { Request } from '../Interfaces';
 import { Transform } from './transforms';
 
-export type FieldToFragmentMapping = {
+export default class ReplaceFieldWithFragment implements Transform {
+  private targetSchema: GraphQLSchema;
+  private mapping: FieldToFragmentMapping;
+
+  constructor(
+    targetSchema: GraphQLSchema,
+    fragments: Array<{
+      field: string;
+      fragment: string;
+    }>,
+  ) {
+    this.targetSchema = targetSchema;
+    this.mapping = {};
+    for (const { field, fragment } of fragments) {
+      const parsedFragment = parseFragmentToInlineFragment(fragment);
+      const actualTypeName = parsedFragment.typeCondition.name.value;
+      this.mapping[actualTypeName] = fragments[actualTypeName] || {};
+      this.mapping[actualTypeName][field] = parsedFragment;
+    }
+  }
+
+  public transformRequest(originalRequest: Request): Request {
+    const document = replaceFieldsWithFragments(
+      this.targetSchema,
+      originalRequest.document,
+      this.mapping,
+    );
+    return {
+      ...originalRequest,
+      document,
+    };
+  }
+}
+
+type FieldToFragmentMapping = {
   [typeName: string]: { [fieldName: string]: InlineFragmentNode };
 };
-
-export default function ReplaceFieldWithFragment(
-  targetSchema: GraphQLSchema,
-  mapping: FieldToFragmentMapping,
-): Transform {
-  return {
-    transformRequest(originalRequest: Request): Request {
-      const document = replaceFieldsWithFragments(
-        targetSchema,
-        originalRequest.document,
-        mapping,
-      );
-      return {
-        ...originalRequest,
-        document,
-      };
-    },
-  };
-}
 
 function replaceFieldsWithFragments(
   targetSchema: GraphQLSchema,
@@ -74,4 +91,31 @@ function replaceFieldsWithFragments(
       },
     }),
   );
+}
+
+function parseFragmentToInlineFragment(
+  definitions: string,
+): InlineFragmentNode {
+  if (definitions.trim().startsWith('fragment')) {
+    const document = parse(definitions);
+    for (const definition of document.definitions) {
+      if (definition.kind === Kind.FRAGMENT_DEFINITION) {
+        return {
+          kind: Kind.INLINE_FRAGMENT,
+          typeCondition: definition.typeCondition,
+          selectionSet: definition.selectionSet,
+        };
+      }
+    }
+  }
+
+  const query = parse(`{${definitions}}`)
+    .definitions[0] as OperationDefinitionNode;
+  for (const selection of query.selectionSet.selections) {
+    if (selection.kind === Kind.INLINE_FRAGMENT) {
+      return selection;
+    }
+  }
+
+  throw new Error('Could not parse fragment');
 }